+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Handle parenthesized and immediately invoked functions in argument lists
   like other function literals (#566).
 * Fix splitting in generic methods with `=>` bodies (#584).
+* Allow splitting between a parameter name and type (#585).
 * Don't split after `<` when a collection is in statement position (#589).
 
 # 0.2.16